requirements.txt
keras~=2.15.0
tensorflow~=2.15.0
numpy~=1.26.2
requests~=2.28.2
autokeras~=1.0.20
Pillow~=10.1.0
asyncio~=3.4.3
discord~=1.7.3
loguru~=0.7.2